% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpCorrResult}
\alias{chimpCorrResult}
\title{Chimpanzee correlations result from TEKRABber}
\format{
An object of class \code{data.frame} with 38400 rows and 5 columns.
}
\usage{
data(chimpCorrResult.rda)
}
\description{
This is a dataframe including the chimpanzee correlation between KZNFs and
TEs pairs using TEKRABber.
}
\examples{
data(chimpCorrResult)
}
\keyword{datasets}
