% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWICE.R
\name{.calculate_TWICE_mixed}
\alias{.calculate_TWICE_mixed}
\title{Internal function for calculating TWiCE scores in 'mixed' mode}
\usage{
.calculate_TWICE_mixed(
  all_NS_res,
  cor_df,
  unique_paths,
  p.adjust.method,
  perm,
  parallel,
  cores,
  min_exist_KO,
  max_exist_KO
)
}
\value{
Data frame with TWiCE scores for mixed mode
}
\description{
Internal function for calculating TWiCE scores in 'mixed' mode
}
\keyword{internal}
