% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWICE.R
\name{.calculate_TWICE_separate}
\alias{.calculate_TWICE_separate}
\title{Internal function for calculating TWiCE scores in 'separate' mode}
\usage{
.calculate_TWICE_separate(
  all_NS_res,
  cor_df,
  unique_paths,
  p.adjust.method,
  spearate_res,
  perm,
  parallel,
  cores,
  min_exist_KO,
  max_exist_KO
)
}
\value{
Data frame with TWiCE scores for separate mode
}
\description{
Internal function for calculating TWiCE scores in 'separate' mode
}
\keyword{internal}
