% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19age}
\alias{hg19age}
\title{hg19 information including age inference}
\format{
An object of class \code{data.frame} with 19462 rows and 3 columns.
}
\usage{
data(hg19age)
}
\description{
This is a dataframe including hg19 gene (Ensembl Ver73), transcript, branch
(0-14), and geneName modified from GenTree data
(http://gentree.ioz.ac.cn/index.php)
}
\examples{
data(hg19age)
}
\keyword{datasets}
