% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19rmsk_gtf}
\alias{hg19rmsk_gtf}
\title{Coordinates information of hg19 repeatmasker reference}
\format{
An object of class \code{data.frame} with 4490853 rows and 5 columns.
}
\usage{
data(hg19rmsk_gtf)
data(hg19rmsk_gtf)
}
\description{
This repeatmasker annotation including the information of transposable
elements and their chromosome number, start, end coordinate position. It
can be used to calculated the distance between KRAB-ZNFs and TEs.
}
\keyword{datasets}
