% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19rmsk_info}
\alias{hg19rmsk_info}
\title{human hg19 repeatmasker annotation file}
\format{
An object of class \code{data.frame} with 992 rows and 3 columns.
}
\usage{
data(hg19rmsk_info)
}
\description{
This is a repeatmasker annotation file including information of transposable
elements in human (hg19). This dataframe object has 3 columns. gene_id for
the name of transposable element, family_id for the class name, and class_id
for the class name.
}
\keyword{datasets}
