% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38age}
\alias{hg38age}
\title{hg38 information including age inference}
\format{
An object of class \code{data.frame} with 18757 rows and 4 columns.
}
\usage{
data(hg38age)
}
\description{
This is a dataframe including hg38 gene (Ensembl Ver95), transcript, branch
(0-14), and geneName modified from GenTree data
(http://gentree.ioz.ac.cn/index.php). There is actual some missing data
when you download them in a tsv file at once from GenTree. If you found NA,
it is suggested to use the search tab to double-check on their website.
}
\examples{
data(hg38age)
}
\keyword{datasets}
