% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmchimpGene}
\alias{hmchimpGene}
\title{Gene Expression data of human and chimpanzee}
\format{
An object of class \code{data.frame} with 6471 rows and 6 columns.
}
\usage{
data(hmchimpGene)
}
\description{
This is a gene expression dataframe including 3 replicates of human and
chimpanzee. Row names are ensembl gene ID and column names are sample ID.
}
\keyword{datasets}
