% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmchimpTE}
\alias{hmchimpTE}
\title{Transposable elements (TE) data of human and chimpanzee}
\format{
An object of class \code{data.frame} with 928 rows and 6 columns.
}
\usage{
data(hmchimpTE)
}
\description{
This is a TE expression dataframe including 3 replicates of human and
chimpanzee. Row names are TE gene ID and column names are sample ID.
}
\keyword{datasets}
