% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmKZNFs337}
\alias{hmKZNFs337}
\title{Human KRAB-Zinc-Finger gene reference data}
\format{
An object of class \code{data.frame} with 337 rows and 2 columns.
}
\usage{
data(hmKZNFs337)
}
\description{
This is a 2 column dataframe including the ensembl gene ID and common gene
name of 337 KRAB-Zinc-Finger gene in human.
}
\keyword{datasets}
