% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KZNFs423}
\alias{KZNFs423}
\title{Information of human 423 KRAB-ZNFs gene}
\format{
An object of class \code{data.frame} with 423 rows and 51 columns.
}
\usage{
data(KZNFs423)
}
\description{
This is a dataframe including 423 human KRAB-ZNFs genes witt coordinates
}
\examples{
data(KZNFs423)
}
\keyword{datasets}
