% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kznfsComplexHeatmap.R
\name{kznfsComplexHeatmap}
\alias{kznfsComplexHeatmap}
\title{Use ComplexHeatmap to visualize KZNFs data from TEKRABber result}
\usage{
kznfsComplexHeatmap(df, top = 40, species = "hsapiens")
}
\arguments{
\item{df}{a TEKRABber geneDESeq2Log2 table}

\item{top}{a numeric value to show top numer of genes in heatmap}

\item{species}{indicate the species name to convert ID}
}
\value{
a ComplexHeatmap heatmap
}
\description{
Use ComplexHeatmap to visualize KZNFs data from TEKRABber result
}
