% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kznfsTEsNetwork.R
\name{kznfsTEsNetwork}
\alias{kznfsTEsNetwork}
\title{Generate KRAB-ZNFs and TEs regulatory network}
\usage{
kznfsTEsNetwork(..., code)
}
\arguments{
\item{...}{add CorrResult Network dataframe from TEKRABber result}

\item{code}{name of your network}
}
\value{
a list including DiffNet, Graph, network, znfName, teName
}
\description{
Use CoDiNA to generate gene regulatory network to combine
KRAB-ZNFs and TEs network between two species.
}
\examples{
data(hmCorrResult)
data(chimpCorrResult)
kznfsTEsNetwork(hmCorrResult, chimpCorrResult, code=c("net1", "net2"))

}
