% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primateGRFs}
\alias{primateGRFs}
\title{Positive selections of primates GRFs}
\format{
An object of class \code{data.frame} with 3215 rows and 30 columns.
}
\usage{
data(primateGRFs)
}
\description{
This dataframe is from (Positive Selection in Gene Regulatory Factors
Suggests Adaptive Pleiotropic Changes During Human Evolution, Jovanovic,
et al. 2021). The data include 27 primates with their gene name, gene ID,
including transcription factors, co-factors and histone-modifying enzymes.
}
\keyword{datasets}
