% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resHmChimpGene}
\alias{resHmChimpGene}
\title{Differentially expressed genes comparing human and chimpanzee}
\format{
An object of class \code{data.frame} with 6471 rows and 5 columns.
}
\usage{
data(resHmChimpGene)
}
\description{
This is a dataframe include log2 differentially expressed genes information
comparing human and chimpanzee from TEKRABber analyzing results.
}
\examples{
data(resHmChimpGene)
}
\keyword{datasets}
