% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resHmChimpTE}
\alias{resHmChimpTE}
\title{Differentially expressed TEs comparing human and chimpanzee}
\format{
An object of class \code{data.frame} with 928 rows and 5 columns.
}
\usage{
data(resHmChimpTE)
}
\description{
This is a dataframe include log2 differentially expressed TEs information
comparing human and chimpanzee from TEKRABber analyzing results.
}
\examples{
data(resHmChimpTE)
}
\keyword{datasets}
