% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmskFilter.R
\name{rmskFilter}
\alias{rmskFilter}
\title{Filter transposable elements from rmsk reference based on given their gene,
family, and class ID.}
\usage{
rmskFilter(rmskInfo, gene = NULL, family = NULL, class = NULL, show = FALSE)
}
\arguments{
\item{rmskInfo}{rmsk dataframe, with three columns, gene_id, family_id, and
class_id}

\item{gene}{a vector including the gene names you want}

\item{family}{a vector including the family names you want}

\item{class}{a vector including the class names you want}

\item{show}{if you want to first see all possible family/class, set this as
TRUE}
}
\value{
a subset of rmsk dataframe including three columns, gene, family and
class
}
\description{
For filtering out the select transposable elements from gene,
family, and class to down-stream analysis, we provide this function to help
quickly list all options out. You can run this function first setting show
to TRUE to see every possibility. Then you provide a character vector
including the name of family/class you want to select. The function will
then return a subset of the rmsk dataframe.
}
\examples{
df <- data(hg19rmsk_info)

#list all possibilities
rmskFilter(df, show=TRUE)

# select family
rmskFilter(df, family="hAT-Charlie")

# select class
df_subset <- rmskFilter(df, class= c("LINE", "SINE"))

}
