% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWICE-package.R
\docType{package}
\name{TWICE-package}
\alias{TWICE}
\alias{TWICE-package}
\title{TWICE: Topology-Weighted Integrated Correlation Enrichment}
\description{
A New Enrichment Method for Exposomics: Topology-Weighted Integrated Correlation Enrichment (TWICE) Analysis
}
\author{
\strong{Maintainer}: Chen Peng \email{bfzede@gmail.com} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

Authors:
\itemize{
  \item Zinuo Huang \email{znhuang@zju.edu.cn}
}

}
\keyword{internal}
