% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWICE.R
\name{.calculate_TWICE_directed}
\alias{.calculate_TWICE_directed}
\title{Internal function for calculating TWiCE scores in 'directed' mode}
\usage{
.calculate_TWICE_directed(
  all_NS_res,
  cor_df,
  unique_paths,
  p.adjust.method,
  perm,
  parallel,
  cores,
  min_exist_KO = 5,
  max_exist_KO = 600
)
}
\value{
Data frame with TWiCE scores for directed mode

Data frame with TWiCE scores for directed mode
}
\description{
Internal function for calculating TWiCE scores in 'directed' mode
}
\keyword{'directed'}
\keyword{Internal}
\keyword{TWiCE}
\keyword{calculating}
\keyword{for}
\keyword{function}
\keyword{in}
\keyword{internal}
\keyword{mode}
\keyword{scores}
