% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensIDtoName.R
\name{ensIDtoName}
\alias{ensIDtoName}
\title{convert the ensID in row to gene name for heatmap visualization}
\usage{
ensIDtoName(df, species)
}
\arguments{
\item{species}{}
}
\value{
a dataframe(matrix) with gene name in row names
}
\description{
convert the ensID in row to gene name for heatmap visualization
}
