% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38age}
\alias{hg38age}
\title{hg38 information including age inference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18705 rows and 3 columns.
}
\usage{
data(hg38age)
}
\description{
This is a dataframe including hg38 gene (Ensembl Ver95), transcript, branch
(0-14), and geneName modified from GenTree data
(http://gentree.ioz.ac.cn/index.php).
}
\examples{
data(hg38age)
}
\keyword{datasets}
