% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmCorrResult}
\alias{hmCorrResult}
\title{Human correlations result from TEKRABber}
\format{
An object of class \code{data.frame} with 38400 rows and 5 columns.
}
\usage{
data(hmCorrResult.rda)
}
\description{
This is a dataframe including the human correlation between KZNFs and TEs
pairsusing TEKRABber.
}
\examples{
data(hmCorrResult)
}
\keyword{datasets}
