% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kznfsTEsPairs.R
\name{kznfsTEsPairs}
\alias{kznfsTEsPairs}
\title{Integrate correlation pairs of KZNFs and TEs}
\usage{
kznfsTEsPairs(df, rmsk, title = NULL, fileName = NULL)
}
\arguments{
\item{fileName}{}
}
\value{
a KZNFs TEs pairs table
}
\description{
Integrate correlation pairs of KZNFs and TEs
}
\examples{


}
