% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teComplexHeatmap.R
\name{teComplexHeatmap}
\alias{teComplexHeatmap}
\title{Use ComplexHeatmap to visualize TEs data from TEKRABber result}
\usage{
teComplexHeatmap(df, teFilter = NULL, top = 40, fileName = NULL)
}
\arguments{
\item{df}{a TEKRABber teDESeq2norm table}

\item{teFilter}{a list of transposable elements subsets}

\item{top}{a numeric value to show top number of genes in heatmap}

\item{fileName}{the name for saving pdf file (if specified)}
}
\value{
a ComplexHeatmap heatmap (Formal class Heatmap)
}
\description{
This function use ComplexHeatmap::Heatmap() to visualize TEs
data generated from using TEKRABber. The input dataframe is the file with
extension teDESeq2norm.csv from TEKRABber. You can specified how many of
the TEs you want to visualize. The default is showing the top 40 TEs
with the highest row means among all samples. You can also specify a list
to filter out a subset of TEs using the teFilter parameters. If you face an
error with could not find function "unit", you need to load the
ComplexHeatmap library to your workspace.
}
\examples{
data(hmchimpTE)
teComplexHeatmap(hmchimpTE, top=40)


}
